Fix ArgumentOutOfRangeException exception when removing a not focused item
The code in RemoveAt doesn't update the focused item index, if the focused
item is the last item in the list and another item is removed.
The next time an item is removed, the focused item index is equal to Count
and an ArgumentOutOfRangeException is thrown.

(cherry picked from commit 59f78905ffd8f223307a49bef3580b22bb00a4fa)