Fix ApplicationSettingsBaseTest:TestSettings1_Properties on Windows.
This is most likely the result of a compiler change. mcs (at least
for old versions) would sort methods by name, with properties
grouped together. Now csc is used which writes methods in the
order they are given in the source. Thus, the Address used to be
first due to the sorting, but now Username is first because it
appears first in the source.

This reveals a Mono bug where System.Configuration on Windows
apparently enumerates methods in the order they appear, while
Mono's System.Configuration lexically sorts them at runtime.