threads: Remove useless priority-setting code.
The "fallback" of setting a priority value based on policy causes
build issues in the case of an unrecognized priority because it
adds output to stdout. Said fallback is, best I can tell, also
completely useless because the value it sets has nothing to do
with the priority that was passed in.

In fact, the whole section only does anything with the priority
value if _POSIX_PRIORITY_SCHEDULING is defined, so it should be
gated on that.