System.Windows.Forms: Fix a handle leak.
According to https://github.com/mono/mono/issues/20769 an DestroyNotify event, which was supposed to delete handles from Hwnd.windows, was skipped and Dispose was supposed to be called for them. Due to the fact that they remained in the Hwnd.windows collection, and in X11 these handles were already considered free , X11 could issue a handle number that is already in Hwnd.windows.