Open the context menu for the control on which the right mouse button is pressed
When WM_CONTEXTMENU arrives in the ListView in an overloaded WndProc, we
expect that with m.wParam != Handle this event comes from the list header,
otherwise it comes from the list itself. Correspondingly when an event
is triggered from the list header, we display the context menu for
the header, otherwise the menu of the list itself will be displayed.
But this event comes from ItemControl and in m.wParam there will be its
Handle, not the list. And the context menu of both the list and the
header will be displayed. Therefore, in ItemControl, when processing the
WM_RBUTTONUP event, we change m.HWnd to owner.Handle, so that the
Handle from ListView goes to HWnd.