When finalizing NativeWindow, delete the handle from Hwnd.windows
If the control is not explicitly destroyed, then the
finalizer tries to destroy it. But the control finalizer does nothing. Next, the
NativeWindow finalizer comes into play, since it has a weak
reference to the control. When the control is explicitly destroyed, the handle
is deleted, but when the control is destroyed by the finalizer, it is not.
It turns out that we must guarantee to remove the handle when finalizing NativeWindow.