arm: Fix assignment of 4-byte float into fpregs.
The original line warns about going outside array bounds, which
is correct - it dereferences random stack memory as well as the
float on the stack. It seems this was a strange way of assigning
the float to the lower 4 bytes of the double, and assigning junk
to the upper 4 bytes.

Oddly, we are using an array of doubles to represent 16 single
floating point registers, 2 of which (if aligned) can hold a
double. This explains the division by 2. But, when assigning
singles to these registers (in add_float), we can use all 16
slots. So for this purpose, we need to treat it as an array
of 16 floats, rather than 8 doubles.