metadata: Fix a build error from the standard bool macro.
This can be defined to _Bool by the standard library. The error
showed up first on macOS, but I think it could happen on any
platform.

We're not really using bool as a type here. We're just using it
for string/identifier building. Undefining the macro seems to be
the most reasonable way to keep it working roughly the same with
the same idenfitiers.

Other files should be able to deal with the bool macro just fine.