System.Windows.Forms: Don't set HRGN on WM_NCPAINT messages.
The HRGN is not properly cleaned up. In order to do so, we would
need to send the message rather than return it from GetMessage,
as Windows does.

However, the HRGN is also unused. Any client wanting to use it
would have to use Win32 APIs to do so, which don't exist on the
platforms where we generate WM_NCPAINT. So this isn't necessary.