nunit-lite: Build with reference assemblies.
Building against Mono's class library means we reference Stack`1
and Queue`1 in mscorlib. In .NET Framework, these types are in
System and have never been in mscorlib. So to make nunitlite.dll
load in .NET Framework, so that the run-test-ondotnet target can
actually work, we must use the reference assemblies.